fixed bug that would set status of user to deleted when a user was either changing their personal details or updating their email or password

+++ user/controller/profile.controller.php
@@ -57,8 +57,8 @@ if(isset($_POST['changePassword'])){
     }
 }
 
-if($_SERVER['REQUEST_METHOD'] === 'POST'){
-    $onlineStatus = 'deleted';
+if($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_POST['online_status'])){
+    $onlineStatus = $_POST['online_status'];
     $delete = $profile->delete($id, $onlineStatus);
     
 }
